update store after mutation
